+++ src/main/java/telegrambot/user/UserSettings.java
@@ -1,4 +1,4 @@
-package com.jc8pg2.telegrambot.user;
+package telegrambot.user;
 
 public class UserSettings {
